pyproject.toml
[tool.poetry]
name = "poetry-goodstart"
version = "0.1.0"
description = ""
authors = ["Anton Braun <merkme@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
uvicorn = "^0.29.0"
loguru = "^0.7.2"
python-telegram-bot = {extras = ["job-queue"], version = "^21.1.1"}
poetry-plugin-export = "^1.7.1"
django-environ= "^0.11.2"
django-asgi-lifespan = "^0.3.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}
django-bootstrap5 = "^24.2"
psycopg2-binary = "^2.9.9"
django-material-admin = "^1.8.6"
django-viewflow = "^2.0.2"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.2"
pre-commit = "^3.7.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.dpoetry ",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "migrations",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 79
indent-width = 4
output-format = "grouped"


[tool.ruff.lint]
select = [
    # pep8-naming
    "N",
    #pyflakes
    "F",
    #pycodestyle
    "E",
    "W",
    #mccabe
    "C90",
    #isort
    "I",
    #pydocstyle
    "D",
    #flake8-bugbear
    "B",
    #flake8-commas
    "COM",
    #flake8-django
    "DJ",
    #flake8-print
    "T20",
    #flake8-pie
    "PIE",
    #flake8-quotes
    "Q",
    #eradicate
    "ERA",
    #flake8-type-checking
    "TCH"

]

ignore = ["D100", "D103", "T201", "D104", "D106", "COM819", "D212", "I001", "DJ001"]

fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
quote-style = "single"

[tool.ruff.lint."flake8-quotes"]
inline-quotes = "single"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"settings*" = ["E501", "F405"]
